Remove the timezone offset from the report!
This caused test failures on devices in DST timezones and made the reports
ugly.

Importantly, this patches the `reporting` module and not the
UploadDescriptor itself, since its serialization is effectively
canonical, and the fact that we lower the struct into json for the
template engine is largely an implementation detail.